+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+- Change All Programs Total key format [#421](https://github.com/policy-design-lab/pdl-frontend/issues/421)
+
 ## [1.11.0] - 2025-07-31
 
 ### Added